requirements.txt
# RetailGPT — GenAI NL-to-SQL
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
